Renaming scil_generate_gradient_sampling.py to scil_gradients_generate_sampling.py
